requirements.txt
aiogram>=3.3.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
transformers>=4.40.0
torch>=2.6.0
safetensors>=0.4.2
accelerate>=0.25.0

